Move NTT logic to a separate module
